Refactored app.js code into dashboard component and set up routes for user auth
