+++ backend-fe/package.json
@@ -8,6 +8,7 @@
     "start": "node server.js",
     "dev": "nodemon sercer.js"
   },
+  "proxy": "http://localhost:4000",
   "keywords": [],
   "author": "",
   "license": "ISC",
